+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 4.3.1
+- Fix Flat Plant Blocks having wood sounds, not Grass Sounds
 
 # 4.3.0
 - Update to 1.21.11, Support for 1.21.10 has now ended